updated the functions

+++ test/TestERC20Facet.sol
@@ -9,7 +9,7 @@ import "../contracts/facets/OwnershipFacet.sol";
 import "../contracts/facets/DiamondLoupeFacet.sol";
 import "../contracts/facets/ERC20Facet.sol";
 
-contract TestERC20Facet is Test, IDiamondCut{
+contract TestERC20Facet is Test, IDiamondCut {
     Diamond private diamond;
     DiamondCutFacet private dCutFacet;
     DiamondLoupeFacet private dLoupe;
@@ -34,25 +34,25 @@ contract TestERC20Facet is Test, IDiamondCut{
 
         cut[0] = (
             FacetCut({
-            facetAddress: address(dLoupe),
-            action: FacetCutAction.Add,
-            functionSelectors: generateSelectors("DiamondLoupeFacet")
-        })
+                facetAddress: address(dLoupe),
+                action: FacetCutAction.Add,
+                functionSelectors: generateSelectors("DiamondLoupeFacet")
+            })
         );
 
         cut[1] = (
             FacetCut({
-            facetAddress: address(ownerF),
-            action: FacetCutAction.Add,
-            functionSelectors: generateSelectors("OwnershipFacet")
-        })
+                facetAddress: address(ownerF),
+                action: FacetCutAction.Add,
+                functionSelectors: generateSelectors("OwnershipFacet")
+            })
         );
         cut[2] = (
             FacetCut({
-            facetAddress: address(erc20Facet),
-            action: FacetCutAction.Add,
-            functionSelectors: generateSelectors("ERC20Facet")
-        })
+                facetAddress: address(erc20Facet),
+                action: FacetCutAction.Add,
+                functionSelectors: generateSelectors("ERC20Facet")
+            })
         );
 
         //upgrade diamond
@@ -68,38 +68,36 @@ contract TestERC20Facet is Test, IDiamondCut{
         diamondErc20 = ERC20Facet(address(diamond));
     }
 
-    function testMintToAddressesBalance() external{
+    function testMintToAddressesBalance() external {
         assertEq(diamondErc20.balanceOf(A), 100_000_000e18);
         assertEq(diamondErc20.balanceOf(B), 100_000_000e18);
     }
 
     function testTransfer() external {
         switchSigner(A);
-        diamondErc20.transfer(address (1), 50_000e18);
+        diamondErc20.transfer(address(1), 50_000e18);
         assertEq(diamondErc20.balanceOf(address(1)), 50_000e18);
         assertEq(diamondErc20.balanceOf(A), 99_950_000e18);
     }
 
-    function testApprove() external{
+    function testApprove() external {
         switchSigner(A);
-        diamondErc20.approve(address (1), 100_000e18);
-        uint result = diamondErc20.allowance(A, address (1));
+        diamondErc20.approve(address(1), 100_000e18);
+        uint256 result = diamondErc20.allowance(A, address(1));
         assertEq(result, 100_000e18);
     }
 
-    function testTransferFrom() external{
+    function testTransferFrom() external {
         switchSigner(A);
-        diamondErc20.approve(address (1), 100_000e18);
-        switchSigner(address (1));
-        diamondErc20.transferFrom(A, address (3), 80_000e18);
+        diamondErc20.approve(address(1), 100_000e18);
+        switchSigner(address(1));
+        diamondErc20.transferFrom(A, address(3), 80_000e18);
         assertEq(diamondErc20.balanceOf(address(3)), 80_000e18);
-        assertEq(diamondErc20.allowance(A,address(1)), 20_000e18);
+        assertEq(diamondErc20.allowance(A, address(1)), 20_000e18);
         assertEq(diamondErc20.balanceOf(A), 99_920_000e18);
     }
 
-    function generateSelectors(
-        string memory _facetName
-    ) internal returns (bytes4[] memory selectors) {
+    function generateSelectors(string memory _facetName) internal returns (bytes4[] memory selectors) {
         string[] memory cmd = new string[](3);
         cmd[0] = "node";
         cmd[1] = "scripts/genSelectors.js";
@@ -109,9 +107,7 @@ contract TestERC20Facet is Test, IDiamondCut{
     }
 
     function mkaddr(string memory name) public returns (address) {
-        address addr = address(
-            uint160(uint256(keccak256(abi.encodePacked(name))))
-        );
+        address addr = address(uint160(uint256(keccak256(abi.encodePacked(name)))));
         vm.label(addr, name);
         return addr;
     }
@@ -126,9 +122,5 @@ contract TestERC20Facet is Test, IDiamondCut{
         }
     }
 
-    function diamondCut(
-        FacetCut[] memory _diamondCut,
-        address _init,
-        bytes calldata _calldata
-    ) external override {}
+    function diamondCut(FacetCut[] memory _diamondCut, address _init, bytes calldata _calldata) external override {}
 }
